+++ src/jupyter_starters/py_starters/cookiecutter.py
@@ -9,7 +9,7 @@ from pathlib import Path
 from tempfile import TemporaryDirectory
 from typing import TYPE_CHECKING, Any, Dict, Text
 
-from notebook.utils import url_path_join as ujoin
+from jupyter_server.utils import url_path_join as ujoin
 
 from ..json_ import JsonSchemaException, json_validator
 from ..types import Status